# docs/swagger.yaml
basePath: /auth
host: lifthus.com
info:
  contact:
    email: lifthus531@gmail.com
    name: API Support
    url: lifthus531@gmail.com
  description: This is Project-Hus's root authentication server containing each user's
    UUID, which is unique for all hus services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Project-Hus auth server
  version: 0.0.0
paths:
  /access:
    get:
      description: |-
        validates the google ID token and redirects with hus refresh token to /auth/{token_string}.
        the access token will be expired in 10 minutes.
      parameters:
      - description: Refresh token
        in: header
        name: jwt
        required: true
        type: string
      responses:
        "201":
          description: Access token created
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: gets refresh token in the header and returns access token in the cookie
        after validation.
      tags:
      - auth
  /google:
    post:
      consumes:
      - application/json
      description: |-
        validates the google ID token and redirects with hus refresh token to /auth/{token_string}.
        the refresh token will be expired in 7 days.
      parameters:
      - description: Google ID token
        in: body
        name: jwt
        required: true
        schema:
          type: string
      responses:
        "301":
          description: to /auth/{token_string}
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: processes google auth and redirect with refresh token in url.
      tags:
      - auth
  /hus:
    post:
      consumes:
      - application/json
      description: |-
        validates the google ID token and redirects with hus refresh token to /auth/{token_string}.
        the refresh token will be expired in 7 days.
      parameters:
      - description: Google ID token
        in: body
        name: jwt
        required: true
        schema:
          type: string
      responses:
        "301":
          description: to /auth/{token_string}
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: processes google auth and redirect with refresh token in url.
      tags:
      - auth
  /refresh:
    get:
      description: |-
        validates the google ID token and redirects with hus refresh token to /auth/{token_string}.
        the access token will be expired in 10 minutes.
      parameters:
      - description: Refresh token
        in: header
        name: jwt
        required: true
        type: string
      responses:
        "201":
          description: Access token created
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      summary: gets refresh token in the header and returns access token in the cookie
        after validation.
      tags:
      - auth
swagger: "2.0"